requirements.txt
absl-py==2.3.1
acres==0.5.0
aiobotocore==2.24.2
aiofiles==23.2.1
aiogram==3.13.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.9.1
aioitertools==0.12.0
aioredis==2.0.1
aiosignal==1.3.1
aiosqlite==0.21.0
aistudio-sdk==0.3.6
annotated-types==0.7.0
anthropic==0.49.0
anyascii==0.3.3
anyio==4.9.0
argcomplete==3.6.2
asttokens==3.0.0
async-timeout==5.0.1
asyncio-throttle==1.0.2
asyncpg==0.30.0
attrs==24.2.0
backports.tarfile==1.2.0
bce-python-sdk==0.9.45
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.37.29
botocore==1.40.18
Brotli==1.1.0
build==1.2.2.post1
CacheControl==0.14.1
cachetools==5.5.0
camelot-py==1.0.9
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
ci-info==0.3.0
cleo==2.1.0
click==8.1.7
cohere==5.14.2
colorama==0.4.6
colorlog==6.9.0
comm==0.2.3
configobj==5.0.9
configparser==7.2.0
crashtest==0.4.1
crewai==0.11.0
cryptography==45.0.6
cssselect==1.2.0
cssselect2==0.8.0
Cython==3.0.11
dataclasses-json==0.6.7
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
distlib==0.3.9
distro==1.9.0
docstring-parser==0.15
dtw-python==1.5.3
dulwich==0.22.7
easyocr==1.7.2
elevenlabs==2.1.0
et_xmlfile==2.0.0
etelemetry==0.3.1
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.115.4
fastavro==1.10.0
fastjsonschema==2.21.1
feedfinder2==0.0.4
feedparser==6.0.11
ffmpeg-python==0.2.0
ffmpy==0.6.0
filelock==3.16.1
Flask==3.1.0
fonttools==4.57.0
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2025.3.2
future==1.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-genai==1.30.0
google-generativeai==0.8.3
google_search_results==2.1.0
googleapis-common-protos==1.65.0
gotrue==2.10.0
gradio==5.32.1
gradio_client==1.10.2
greenlet==3.1.1
griffe==1.7.2
groovy==0.1.2
groq==0.22.0
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
h2==4.1.0
h5py==3.14.0
hpack==4.0.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
hyperframe==6.0.1
idna==3.10
imageio==2.37.0
imagesize==1.4.1
importlib_metadata==8.4.0
importlib_resources==6.5.2
installer==0.7.0
instructor==0.5.2
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.5
jiter==0.7.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.8.1
keyring==25.6.0
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-openai==0.0.2.post1
langchain-text-splitters==0.3.8
langdetect==1.0.9
langextract==1.0.6
langgraph==0.0.15
langsmith==0.3.45
lazy_loader==0.4
logfire-api==3.12.0
looseversion==1.3.0
lxml==5.3.0
lxml_html_clean==0.4.1
magic-filter==1.0.12
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.22.0
matplotlib-inline==0.1.7
mcp==1.6.0
mdurl==0.1.2
mistralai==1.6.0
ml_collections==1.1.0
modelscope==1.29.1
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.5
newspaper3k==0.2.8
nibabel==5.3.2
ninja==1.13.0
nipype==1.10.0
nltk==3.9.1
nsepython==2.94
numpy==2.2.6
oauthlib==3.2.2
onnx==1.18.0
openai==1.71.0
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opt-einsum==3.3.0
orjson==3.10.18
packaging==25.0
paddleocr==3.2.0
paddlepaddle==3.1.1
paddlex==3.2.1
pandas==2.2.3
parso==0.8.5
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.0.0
pinecone-client==5.0.1
pinecone-plugin-assistant==1.1.0
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
pkginfo==1.12.0
platformdirs==4.3.6
poetry==2.0.0
poetry-core==2.0.0
postgrest==0.18.0
prettytable==3.16.0
prompt_toolkit==3.0.50
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
prov==2.1.1
psutil==7.0.0
psycopg2==2.9.10
pure_eval==0.2.3
puremagic==1.30
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.8.2
pydantic-ai==0.0.53
pydantic-ai-slim==0.0.53
pydantic-evals==0.0.53
pydantic-graph==0.0.53
pydantic-settings==2.7.1
pydantic_core==2.20.1
pydot==4.0.1
pydub==0.25.1
pydyf==0.11.0
pyenchant==3.2.2
Pygments==2.18.0
PyJWT==2.10.1
pym==0.1.3
PyMuPDF==1.26.4
pync==2.0.3
pynput==1.8.1
pyotp==2.9.0
pyparsing==3.2.0
pypdf==3.17.4
pypdfium2==4.30.0
pyphen==0.17.2
pypiwin32==223
pyproject_hooks==1.2.0
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-doctr==1.0.0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
pywin32==310
pywin32-ctypes==0.2.3
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==27.0.2
RapidFuzz==3.11.0
razorpay==1.4.2
rdflib==7.2.1
realtime==2.0.6
redis==6.2.0
regex==2023.12.25
reportlab==4.3.1
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
ruff==0.11.12
s3transfer==0.11.4
safehttpx==0.1.6
safetensors==0.5.3
schedule==1.2.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.14.1
semantic-version==2.10.0
sgmllib3k==1.0.0
shapely==2.1.1
shellingham==1.5.4
simplejson==3.20.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.41.3
storage3==0.9.0
structlog==25.4.0
supabase==2.10.0
supafunc==0.7.0
sympy==1.14.0
tabula==1.0.5
tabula-py==2.10.0
tabulate==0.9.0
tavily-python==0.3.1
tenacity==8.5.0
threadpoolctl==3.6.0
tifffile==2025.8.28
tiktoken==0.5.2
tinycss2==1.4.0
tinyhtml5==2.0.0
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.0
traitlets==5.14.3
traits==7.0.2
transformers==4.52.4
trove-classifiers==2025.1.6.15
twilio==9.4.4
typer==0.16.0
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.11.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
validators==0.35.0
virtualenv==20.28.1
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
websockets==13.1
Werkzeug==3.1.3
win10toast==0.9
wrapt==1.16.0
yarl==1.17.1
zipp==3.20.2
zopfli==0.2.3.post1
zstandard==0.23.0
